+++ mockup/js/config.js
@@ -80,6 +80,7 @@
       'mockup-patterns-querystring': 'patterns/querystring/pattern',
       'mockup-patterns-relateditems': 'patterns/relateditems/pattern',
       'mockup-patterns-relateditems-upload': 'patterns/relateditems/upload',
+      'mockup-patterns-relateditems-url': 'patterns/relateditems',
       'mockup-patterns-recurrence': 'patterns/recurrence/pattern',
       'mockup-patterns-resourceregistry': 'patterns/resourceregistry/pattern',
       'mockup-patterns-resourceregistry-url': 'patterns/resourceregistry',